cv-chatbot/requirements.txt
sentence-transformers
faiss-cpu
gradio
langchain
huggingface-hub
mistral-inference
transformers
torch                       
huggingface-hub
protobuf
sentencepiece
accelerate